==> real_data/02_TCGA_mediation.Rmd <==
---
title: "Survival analysis with deconvolution exposition"
output:
  html_document:
    toc: true
    theme: united
date: "Last compiled on `r format(Sys.time(), '%d %B, %Y')`"
---

High-dimensional mediation analysis: identification of potential DNA methylation probes mediating the pathway between tobacco consumption and survival in PDAC patients.



```{r med_load}
# TCGA data
library(survival)
library(ggplot2)
library(see)
library(dplyr)
library(UpSetR)

tcga_data = readRDS("results/01_tcga_data_expo_deconv.rds")

# define variables
time = tcga_data$time
status = tcga_data$status
age = tcga_data$age
gender = tcga_data$gender
grade = tcga_data$grade
X_deconv = tcga_data$X_deconv$Consensus
M = tcga_data$M
tobacco_bin = tcga_data$tobacco_bin
#alcohol = tcga_data$alcohol

# keep only immune cells
immune_cells = c("Neutrophils","Macrophages","B cells", "T cells", "NK","DCs")
X_immune = X_deconv[rownames(X_deconv) %in% immune_cells,]
```


# Mediation analysis

## Tobacco -> DNAm

```{r med_run}
#mediation tobacco_dnam_survival with immune cells, LF estimation integration K = 8
source("surv_parametric_alternative.R")

exposure = as.numeric(tobacco_bin)
survival_time = as.vector(time)
censoring_status = as.vector(status)
#M = as.matrix(M_sample_13)
M_met = as.matrix(M)
K = 8
immune_cells = t(X_immune[-1,])

covar_U = as.matrix(cbind(exposure, immune_cells,tcga_data$grade, tcga_data$age,tcga_data$gender))
covar = as.matrix(cbind(tcga_data$grade, tcga_data$age,tcga_data$gender))
######################
# mediation with parametric model (survreg function)
######################
hdmax2_step1_param = run_AS_surv_param_alter(exposure, survival_time, censoring_status, covar_U = covar_U, covar= covar, M_met, K = K)
m_param = M_met[, names(sort(hdmax2_step1_param$max2_pvalues))[1:10]]
hdmax2_step2_param = estimate_effect_surv_param(hdmax2_step1_param, m = m_param)
hdmax2_param_tobacco_dnam = list(hdmax2_step1_param=hdmax2_step1_param,
                                 m_param=m_param,
                                 hdmax2_step2_param=hdmax2_step2_param)
saveRDS(hdmax2_param_tobacco_dnam, "results/02_tcga_med_tobacco_dnam_V2_K8_corrected.rds")

```
