==> real_data/03_TCGA_survival_deconv_tobacco_AMR.Rmd <==
---
title: "AMR search + Mediation with AMR Data"
output:
  html_document:
    toc: true
    theme: united
date: "Last compiled on `r format(Sys.time(), '%d %B, %Y')`"
---


Identify Aggregated Methylated Regions (AMR) with `AMR_search` function which uses the P-values from the max-squared test computed with the `run_AS` function and using a adapted comb-p method (comb-p is a tool that manipulates BED files of possibly irregularly spaced P-values and calculates auto-correlation, combines adjacent P-values, performs false discovery adjustment, finds regions of enrichment and assigns significance to those regions). AMR identification could be useful in EWAS when single CpG mediation is unsuccessful, or could be a complementary analysis.

```{r amr_init}
library(survival)
library(ggplot2)
library(epimedtools)
source("surv_parametric_alternative.R")
source("helper_functions.R")

source("surv_aalen_alternative.R")
set.seed(123)
```



```{r amr_load4}
# TCGA data V2 K8 corrected
med_tobacco_dnam = readRDS("results/02_tcga_med_tobacco_dnam_V2_K8_corrected.rds")
data_tcga = readRDS("results/01_tcga_data_expo_deconv.rds")

```

# TCGA AMR mediation

```{r amr_tobacco DNAm AMR V2 K8, message=FALSE, eval=TRUE}

### Tobacco -> DNAm
### 2 cpg min
### seed = 0.05

set.seed(123)
######################
#AMR research
######################
pf_meth = data.frame("Chromosome"=data_tcga$probes_info$CHR,
                     "Start"=data_tcga$probes_info$BP,
                     "End"=data_tcga$probes_info$End,
                     "Gene_Symbol"=data_tcga$probes_info$Gene_Symbol)
rownames(pf_meth) = colnames(data_tcga$M)
saveRDS(pf_meth, "results/03_platform_info_tcga_meth.rds")
# AMR research
res.amr_search = AMR_search(chr = pf_meth$Chromosome,
                                    start = pf_meth$Start,
                                    end = pf_meth$End,
                                    pval = med_tobacco_dnam$hdmax2_step1_param$max2_pvalues,
                                    cpg = rownames(pf_meth),
                                    seed = 0.05, 
                                    nCores = 1)

res.amr_build = AMR_build(res.amr_search,
                                  methylation = data_tcga$M, 
                                  nb_cpg = 2)

# total AMR
m = as.matrix(res.amr_build$AMR_mean)

saveRDS(m, "results/03_tcga_mean_meth_AMR_fdr0_05_V2_K8_corrected.rds")
saveRDS(res.amr_build, "results/03_tcga_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")


# top 50 AMR (most significative according to combp pval)
res = res.amr_build$res
mean_meth = res.amr_build$AMR_mean
cpg_related = res.amr_build$CpG_for_each_AMR
order_AMR = res[order(res.amr_build$res$p),]
top_50_AMR = order_AMR[1:50,]
mean_meth_50 = mean_meth[,top_50_AMR$AMR]
cpg_related_50 = cpg_related[top_50_AMR$AMR]
rownames(mean_meth_50) = rownames(data_tcga$M)


rownames(m) = rownames(data_tcga$M)
mean_meth_50 = as.matrix(mean_meth_50)

## selected mediators effects estimation
hdmax2_step2 = estimate_effect_surv_param(object = med_tobacco_dnam$hdmax2_step1_param, m = mean_meth_50)
acme = hdmax2_step2$effects$univariate$ACME
significative_AMR_acme = acme[which(acme$pval<=0.05),]
mean_meth_significative = m[, significative_AMR_acme$feat]
saveRDS(mean_meth_significative, "results/03_tcga_significative_from_top50_mean_meth_AMR_fdr0_05_V2_K8_corrected.rds")

AMR_info = res.amr_build$res
CpG_related_info = res.amr_build$CpG_for_each_AMR
CpG_related_info_significative = res.amr_build$CpG_for_each_AMR[AMR_info$AMR %in% significative_AMR_acme$feat]
significative_AMR_info = AMR_info[AMR_info$AMR %in% significative_AMR_acme$feat,]

res_50 = list(all_step2_res = hdmax2_step2,
              step2_res_50 = acme,
              AMR_info_50  = AMR_info[AMR_info$AMR %in% top_50_AMR$AMR,],
              cpg_related_50 = cpg_related_50,
              mean_meth_mat = mean_meth_50)
saveRDS(res_50, "results/03_tcga_top50_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")

res = list(
  step2_res = significative_AMR_acme,
  AMR_info = significative_AMR_info,
  CpG_related_info = CpG_related_info_significative
)

saveRDS(res, "results/03_tcga_significative_from_top50_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")

# for causal direction analyses (Nelle)
write.csv(mean_meth_50, "results/03_tcga_AMR_mean_meth_top50_fdr0_05_V2_K8_corrected.csv")

write.csv(mean_meth_significative, "results/03_tcga_AMR_mean_meth_significative_fdr0_05_V2_K8_corrected.csv")

X_deconv = data_tcga$X_deconv$Consensus
immune_cells = c("Macrophages","B cells", "T cells", "NK","DCs")
X_immune = X_deconv[rownames(X_deconv) %in% immune_cells,]
X_immune = t(X_immune)

write.csv(X_immune, "results/03_tcga_consensus_deconv_immune_cells.csv")
```
