==> real_data/04_TCGA_immune_AMR_correlation.Rmd <==
---
title: "Correlation between AMR and Immune cells"
output:
  html_document:
    toc: true
    theme: united
date: "Last compiled on `r format(Sys.time(), '%d %B, %Y')`"
---

```{r corr_AMR immune V2 K8}
library(reshape2)

deconv = readRDS("results/01_tcga_data_expo_deconv.rds")
mean_meth_AMR = readRDS("results/03_tcga_top50_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")
mean_meth_AMR = mean_meth_AMR$mean_meth_mat
X_deconv = deconv$X_deconv$Consensus
immune_cells = c("Macrophages","B cells", "T cells", "NK","DCs")
X_immune = X_deconv[rownames(X_deconv) %in% immune_cells,]
immune_cons = t(X_immune)

dim(mean_meth_AMR)
dim(immune_cons)

cor_matrix <- cor(mean_meth_AMR, immune_cons, use = "pairwise.complete.obs", method = "pearson")

# Make a df
cor_df <- as.data.frame(as.table(cor_matrix))
colnames(cor_df) <- c("DNAm", "Immune", "Correlation")

get_pval <- function(x, y) {
  test <- cor.test(x, y, use = "pairwise.complete.obs", method = "pearson")
  return(test$p.value)
}

get_corr <- function(x, y) {
  test <- cor.test(x, y, use = "pairwise.complete.obs", method = "pearson")
  return(test$estimate)
}

pval_matrix <- matrix(NA, ncol = ncol(immune_cons), nrow = ncol(mean_meth_AMR),
                      dimnames = list(colnames(mean_meth_AMR),
                                      colnames(immune_cons)))
for (i in 1:ncol(mean_meth_AMR)) {
  for (j in 1:ncol(immune_cons)) {
    pval_matrix[i, j] <- get_pval(mean_meth_AMR[, i], immune_cons[, j])
  }
}

corr_matrix <- matrix(NA, ncol = ncol(immune_cons), nrow = ncol(mean_meth_AMR),
                      dimnames = list(colnames(mean_meth_AMR),
                                      colnames(immune_cons)))
for (i in 1:ncol(mean_meth_AMR)) {
  for (j in 1:ncol(immune_cons)) {
    corr_matrix[i, j] <- get_corr(mean_meth_AMR[, i], immune_cons[, j])
  }
}

# Long format
cor_df <- melt(cor_df, varnames = c("DNAm", "Immune"), value.name = "Correlation")
pval_df <- melt(pval_matrix, varnames = c("DNAm", "Immune"), value.name = "P_value")
corr_df <- melt(corr_matrix, varnames = c("DNAm", "Immune"), value.name = "corr")
pval_df$variable = rep("P_value", dim(pval_df)[1] )

# plot_df = data.frame(AMR = cor_df$DNAm,
#                      correlation = cor_df$Correlation,
#                      pvalue = pval_df$P_value,
#                      cell_type = pval_df$Immune,
#                      corr = corr_df$corr)
df_to_save = data.frame(AMR = cor_df$DNAm,
                        pvalue = pval_df$P_value,
                        cell_type = pval_df$Immune,
                        corr = corr_df$corr)
# correlation value, pvalue
saveRDS(df_to_save, "results/04_tcga_AMR_imm_corr_fdr0_05_V2_K8_corrected.rds")

# AMR mean methylation matrix + immune cell types proportion
res = list(immune_cons = immune_cons, mean_meth_AMR = mean_meth_AMR)
saveRDS(res, "results/04_tcga_imm_AMR_matrix_fdr0_05_V2_K8_corrected.rds")
```
