==> real_data/05_causality.Rmd <==
---
title: "Causality analysis"
output:
  html_document:
    toc: true
    theme: united
date: "Last compiled on `r format(Sys.time(), '%d %B, %Y')`"
---


Causal discovery framework based on conditional independence tests to identify causal relationships among the following variables: tobacco exposure, mediating AMR, total immune infiltration, B cell infiltration proportion, dendritic cell infiltration proportion, and survival. 


```{r caus_load}
# load packages
library(ggplot2)   
#install.packages("mpmi")
library(mpmi)

# For PC analysis
library(reticulate)
library(tidyverse)
library(ggpubr)
library(dagitty)
library(broom)
library(pheatmap)
# load data

library(survival)
library(survminer)
library(gridExtra)

# for theoretical exploration
AMR_matrix = readRDS("results/04_tcga_imm_AMR_matrix_fdr0_05_V2_K8_corrected.rds")
tcga_data = readRDS("tcga_pdac_mediation/results/01_tcga_data_expo_deconv.rds")

# for independance testing
datAMR = read.csv2("results/03_tcga_AMR_mean_meth_top50_fdr0_05_V2_K8_corrected.csv", header = TRUE, stringsAsFactors = FALSE, row.names = 1, sep = ",") #to check

#significant AMR
tobacco_AMR <- readRDS("results/03_tcga_significative_from_top50_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")

datIMM = read.csv2("results/03_tcga_consensus_deconv_immune_cells.csv", header = TRUE, stringsAsFactors = FALSE, row.names = 1, sep = ",")
datIMM[] <- lapply(datIMM, function(x) as.numeric(as.character(x)))
datIMM$all = rowSums(datIMM)

#result of the mediation to collect latent factors
res_med = readRDS("results/02_tcga_med_tobacco_dnam_V2_K8_corrected.rds")
LFs = res_med$hdmax2_step1_param$AS_1$U
colnames(LFs) = paste0("LF_", c("A", "B", "C", "D", "E", "F", "G", "H"))

#Match patient
patient_id = rownames(tcga_data$M)
datIMM = datIMM[patient_id,]

```



## A generic function with mutual information

```{r}
function_causality = function(x1, x2){
  # model 1
  mod_forward = lm(x2 ~ x1)
  res_forward = mod_forward$residuals
   mi_cont <- mpmi::cmi(cbind(x1,res_forward)) # mutual information on continuous variables
    zscore_fw = lapply(mi_cont, round, 2)$zvalues[1,2] #Z-scores for each hypothesis that the corresponding BCMI value is zero. These have poor statistical properties but can be useful as a rough measure of the strength of association.
  #model 2
      mod_backward = lm(x1 ~ x2)
  res_backward = mod_backward$residuals
  mi_cont <- mpmi::cmi(cbind(x2,res_backward)) # mutual information on continuous variables
    zscore_bw = lapply(mi_cont, round, 2)$zvalues[1,2] #Z-scores for each hypothesis that the corresponding BCMI value is zero. These have poor statistical properties but can be useful as a rough measure of the strength of association.
  return(list(x1 = x1, x2 = x2, forward = res_forward, backward = res_backward, zscore_fw = zscore_fw, zscore_bw = zscore_bw))
}


```

## Application on real data 

```{r}
imm_tot = tcga_data$prop_immune

plot(density(imm_tot))

cor_matrix = apply(AMR_matrix$mean_meth_AMR, 2, FUN = function(AMR) { 
                   cor =cor.test(AMR,imm_tot)$estimate
                   pval = cor.test(AMR,imm_tot)$p.value
                   return(list(cor = cor, pvalue = pval))
                     })

cor_matrix = do.call(rbind,cor_matrix)
cor_matrix = data.frame(lapply(data.frame(cor_matrix, stringsAsFactors=FALSE), unlist), stringsAsFactors=FALSE)
rownames(cor_matrix) = colnames(AMR_matrix$mean_meth_AMR)

AMR_subset = rownames(cor_matrix) 
length(AMR_subset)
AMR_subset

res_causality = apply(AMR_matrix$mean_meth_AMR[,AMR_subset],2, function(x) {
  function_causality(x,imm_tot)
  })


#save results
saveRDS(res_causality, "results/05_tcga_causality.rds")
```
 

 

```{r}
datAMR[] <- lapply(datAMR, function(x) as.numeric(as.character(x)))
colnames(datIMM) = c( "Macrophages",  "B_cells",     "T_cells" ,    "NK",          "DCs" ,        "all" )

smoking = ifelse(tcga_data$tobacco==0, 'Non-smoker', 'Smoker')
names(smoking) = rownames(tcga_data$M)
labels = smoking[rownames(datAMR)]


```


# Analysis for the paper

## Test and plot Immune fraction / survival 

```{r}

 pval_thres = 0.15

 #Select Imm correlated with survival
 colnames(datIMM)
 colnames(datIMM) = c("Macrophages", "B cells", "T cells", "NK", "DCs", "Tot. Imm." )
  res = apply(datIMM, 2, function(x) {
   model = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ x + tcga_data$age + tcga_data$gender +  tcga_data$grade)
    summary(model)$coefficients[, "Pr(>|z|)"]
 })
  

 immune_names = names(which(res[1,] <= pval_thres)) # select imm -> survival in the graph
 length(immune_names)
 
  #Plot Kaplan Meyer curves

selected_vars <- colnames(datIMM)[1:6]  
```


## Run causal discovery

```{r}

# Data processing
AMR_names = colnames(datAMR)
smoking = ifelse(tcga_data$tobacco==0, 'Non-smoker', 'Smoker')
names(smoking) = rownames(tcga_data$M)
labels = smoking[rownames(datAMR)]
smoking_status = as.numeric(as.factor(labels))

pval_thres = 0.15

# details of each model for publication:

prop100= datIMM[,"Tot. Imm."]*100 #to interpret the HR has increase of 1 unit (1% of immune infiltration) lead to HR likely to dire than persons with 1% less
mod = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ prop100 + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(mod)

prop100= datIMM[,"B cells"]*100 #to interpret the HR has increase of 1 unit (1% of immune infiltration) lead to HR likely to dire than persons with 1% less
mod = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ prop100 + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(mod)

prop100= datIMM[,"T cells"]*100 #to interpret the HR has increase of 1 unit (1% of immune infiltration) lead to HR likely to dire than persons with 1% less
mod = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ prop100 + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(mod)

# Latent factor assessment

cor = apply(LFs, 2, function(x) {
  cor.test(x, smoking_status)$p.value
})




# Step 1: Unconditional independence resting

## Tobacco-Survival association: Cox proportional hazards model

SURV_TOB = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~  smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(SURV_TOB)

## Tobacco-Immune associations: Linear regression models

IMMtot_TOB = lm(datIMM[ ,"Tot. Imm."]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade )    
summary(IMMtot_TOB) 

IMMbcell_TOB = lm(datIMM[ ,"B cells"]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade )    
summary(IMMbcell_TOB) 

IMMDCs_TOB = lm(datIMM[ ,"T cells"]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(IMMDCs_TOB) 


AMR_names = tobacco_AMR$AMR_info$AMR

num_res = list()

for (AMR in AMR_names){ 
  
  for (imm in immune_names) {
   
    
       pair = paste(AMR, "-", imm, sep = "")
 print(paste0("Testing AMR:", AMR, " and Immune:", imm))
 
 # Check consisty of T -> AMR -> S path
  
 #  Check the link between tobacco and the AMR 
    mod_tob = lm(datAMR[,AMR]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade)
   tob_sign =  summary(mod_tob)$coefficients[2,4]  < pval_thres 
   
    #  check the link between the AMR and survival
    mod_AMR = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~  datAMR[,AMR] + tcga_data$age + tcga_data$gender +  tcga_data$grade)
   AMR_sign =  summary(mod_AMR)$coefficients[1, "Pr(>|z|)"] < pval_thres 
   
   # Step 2: Conditional independence testing
 
   if (tob_sign == TRUE & AMR_sign == TRUE) {
  
    print(paste0(" AMR:", AMR, " and Immune:", imm, " is kept for doanwstream analysis. "))
   # (Model 1) Which node coefficient will lose its significance between AMR and imm ?
    mod_part = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~  datIMM[ ,imm] + datAMR[,AMR]  + tcga_data$age + tcga_data$gender +  tcga_data$grade)
    summary(mod_part)
    part_surv_sign = summary(mod_part)$coefficients[1:2, "Pr(>|z|)"] 

   
    # (Model 2) check the link between the immune and smoking status when controlling for AMR
    dat = data.frame(
      smoking_status = as.numeric(as.factor(labels)) -1,
             AMR = datAMR[,AMR],
      imm = datIMM[ ,imm],
      age = tcga_data$age,
                 gender = tcga_data$gender,
                 grade =   tcga_data$grade
    )
    
    mod = glm(smoking_status~ ., data = dat, family = binomial(link = "logit"))
    summary(mod)
    part_imm_sign = summary(mod)$coefficients[2:3,4] # check if imm is significant
    
    num_res[[pair]] = c( part_surv_sign,
                        part_imm_sign)
    names(num_res[[pair]]) = c("part_IMM_sign_to_surv", 
                               "part_AMR_sign_to_surv_when_IMM_controlled",
                               "part_AMR_sign_to_smoking",
                               "part_IMM_sign_to_smoking_when_AMR_controlled")
   } else {
     print(paste0(" AMR:", AMR, " and Immune:", imm, " is not kept for doanwstream analysis. "))
     #num_res[[pair]] = c(NA, NA, NA, NA)
     #names(num_res[[pair]]) = c("part_IMM_sign_to_surv", 
    #                            "part_AMR_sign_to_surv_when_IMM_controlled",
     #                           "part_AMR_sign_to_smoking",
     #                           "part_IMM_sign_to_smoking_when_AMR_controlled")
   }

  }


}
num_res = do.call(rbind, num_res)
num_res=data.frame(num_res)

# Directed acyclic graph 1
res_CAT1 = num_res[  num_res$part_IMM_sign_to_surv < pval_thres &
            num_res$part_AMR_sign_to_surv_when_IMM_controlled < pval_thres, ]
res_CAT1
rownames(res_CAT1)
saveRDS(rownames(res_CAT1), file = "results/05_DAG_validated_CAT1.rds")

# Directed acyclic graph 2
res_CAT2 = num_res[  num_res$part_IMM_sign_to_surv < pval_thres &
            num_res$part_AMR_sign_to_surv_when_IMM_controlled >= pval_thres, ]
rownames(res_CAT2)
saveRDS(rownames(res_CAT2), file = "results/05_DAG_validated_CAT2.rds")





```


## Run causal discovery with LF

### Correlation

```{r}

# Compute Pearson correlation matrix between LFs and datIMM
cor_matrix_LF <- cor(LFs, datIMM, use = "pairwise.complete.obs", method = "pearson")

# Function to get p-value from Pearson correlation test between two vectors
get_pval <- function(x, y) cor.test(x, y, use = "pairwise.complete.obs", method = "pearson")$p.value

# Vectorized computation of p-values for all combinations of columns in LFs and datIMM
pval_matrix_LF <- outer(
  1:ncol(LFs), 1:ncol(datIMM),
  Vectorize(function(i, j) get_pval(LFs[, i], datIMM[, j]))
)

# Set row and column names of the p-value matrix to match LFs and datIMM column names
dimnames(pval_matrix_LF) <- list(colnames(LFs), colnames(datIMM))

# Bonferroni adjusted significance threshold
alpha_adj <- 0.01 / (50+ 6) # 50 LFs + 6 immune variables



# Compute Pearson correlation matrix between LFs and datAMR
cor_matrix_AMR <- cor(LFs, datAMR, use = "pairwise.complete.obs", method = "pearson")

# Function to get p-value from Pearson correlation test between two vectors
get_pval <- function(x, y) cor.test(x, y, use = "pairwise.complete.obs", method = "pearson")$p.value

# Vectorized computation of p-values for all combinations of columns in LFs and datAMR
pval_matrix_AMR <- outer(
  1:ncol(LFs), 1:ncol(datAMR),
  Vectorize(function(i, j) get_pval(LFs[, i], datAMR[, j]))
)

# Set row and column names of the p-value matrix to match LFs and datAMR column names
dimnames(pval_matrix_AMR) <- list(colnames(LFs), colnames(datAMR))

# Bonferroni adjusted significance threshold
alpha_adj <- 0.01 / (50+ 6) # 50 LFs + 6 immune variables

# Convert correlation matrix to a data frame for ggplot

#### SELECT COMBINATIONS :

# Vecteurs de noms
immune_vars <- colnames(datIMM)
amr_vars <- colnames(datAMR)
lf_vars <- colnames(LFs)

# Liste pour stocker les résultats
pairs_list <- list()

# Pour chaque combinaison Immune–AMR
for (immune in immune_vars) {
  for (amr in amr_vars) {
    
    # Récupère les corrélations LF vs immune et LF vs amr
    #cor_immune <- cor_matrix_LF[, immune]
    #cor_amr <- cor_matrix_AMR[, amr]
    pval_immune <- pval_matrix_LF[, immune]
    pval_amr <- pval_matrix_AMR[, amr]
    
    # Sélectionne les LFs où les deux corrélations sont > 0.4 en valeur absolue
    #valid_lfs <- lf_vars[abs(cor_immune) > 0.4 & abs(cor_amr) > 0.4]
    valid_lfs <- lf_vars[pval_immune <= alpha_adj & pval_amr <= alpha_adj]
    
    # S'il y a des LFs valides, les stocker
    if (length(valid_lfs) > 0) {
      pairs_list[[paste( amr,immune, sep = "-")]] <- valid_lfs } 
    else {
      pairs_list[[paste( amr,immune, sep = "-")]] <- NA
    } 
  }
}

# Affiche un exemple
pairs_list[1]

saveRDS(pairs_list, file = "results/05_signLFs_by_pairs-A-I.rds")

```

### Causal discovery

```{r}

# Data processing
AMR_names = colnames(datAMR)
smoking = ifelse(tcga_data$tobacco==0, 'Non-smoker', 'Smoker')
names(smoking) = rownames(tcga_data$M)
labels = smoking[rownames(datAMR)]
smoking_status = as.numeric(as.factor(labels))

pval_thres = 0.1

# details of each model for publication:

prop100= datIMM[,"Tot. Imm."]*100 #to interpret the HR has increase of 1 unit (1% of immune infiltration) lead to HR likely to dire than persons with 1% less
mod = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ prop100 + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(mod)

prop100= datIMM[,"B cells"]*100 #to interpret the HR has increase of 1 unit (1% of immune infiltration) lead to HR likely to dire than persons with 1% less
mod = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ prop100 + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(mod)

prop100= datIMM[,"DCs"]*100 #to interpret the HR has increase of 1 unit (1% of immune infiltration) lead to HR likely to dire than persons with 1% less
mod = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ prop100 + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(mod)

# Latent factor assessment

surv_LF = apply(LFs, 2, function(x) {
   survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~ x)
})

surv_LF

cor = apply(LFs, 2, function(x) {
  cor.test(x, smoking_status)$p.value
})


# Step 1: Unconditional independence resting

## Tobacco-Survival association: Cox proportional hazards model

SURV_TOB = survival::coxph(survival::Surv(tcga_data$time, tcga_data$status) ~  smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade )
summary(SURV_TOB)

## Tobacco-Immune associations: Linear regression models

IMMtot_TOB = lm(datIMM[ ,"Tot. Imm."]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade )    
summary(IMMtot_TOB) 

IMMbcell_TOB = lm(datIMM[ ,"B cells"]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade )    
summary(IMMbcell_TOB) 

IMMDCs_TOB = lm(datIMM[ ,"DCs"]~smoking_status + tcga_data$age + tcga_data$gender +  tcga_data$grade)
summary(IMMDCs_TOB) 


AMR_names = tobacco_AMR$AMR_info$AMR

num_res = list()

for (AMR in AMR_names){ 
  
  for (imm in immune_names) {
   
    
       pair = paste(AMR, "-", imm, sep = "")
 print(paste0("Testing AMR:", AMR, " and Immune:", imm))
 
 # Get significant LFs 
  LFs_names = pairs_list[[pair]]
  covar =   data.frame(age = tcga_data$age,
                         gender = tcga_data$gender,
                         grade =   tcga_data$grade
    )
  if (is.na(LFs_names[1])) {
    print(paste0("No significant LFs for pair: ", pair))
  } else {
    print(paste0("Significant LFs for pair: ", pair, " are: ", paste(LFs_names, collapse = ", ")))
    covar = cbind(covar, LFs[, LFs_names, drop = FALSE])
  }
  

 # Check consisty of T -> AMR -> S path
  
 #  Check the link between tobacco and the AMR 
  
  df_model <- data.frame(
  y = datAMR[, AMR],
  smoking_status = smoking_status,
  covar  
)
  mod_tob <- lm(y ~ ., data = df_model)
     tob_sign =  summary(mod_tob)$coefficients[2,4]  < pval_thres 
   
    #  check the link between the AMR and survival
     
     df_model <- data.frame(
  y = datAMR[, AMR],
  time = tcga_data$time,
  status = tcga_data$status,
  covar)
     
     mod_AMR <- survival::coxph(
  survival::Surv(time, status) ~ y +  ., 
  data = df_model  
)

   AMR_sign =  summary(mod_AMR)$coefficients[1, "Pr(>|z|)"] < pval_thres 
   
   # Step 2: Conditional independence testing
 
   if (tob_sign == TRUE & AMR_sign == TRUE) {
  
    print(paste0(" AMR:", AMR, " and Immune:", imm, " is kept for doanwstream analysis. "))
   # (Model 1) Which node coefficient will lose its significance between AMR and imm ?
     
     df_model <- data.frame(
  time = tcga_data$time,
  status = tcga_data$status,
  imm = datIMM[[imm]],
  amr = datAMR[[AMR]],
   covar
)
     mod_part <- survival::coxph(
  survival::Surv(time, status) ~ imm + amr +  .,
  data = df_model
)
    part_surv_sign = summary(mod_part)$coefficients[1:2, "Pr(>|z|)"] 

   
    # (Model 2) check the link between the immune and smoking status when controlling for AMR
    
    
     df_model <- data.frame(
   smoking_status = as.numeric(as.factor(labels)) -1,
             AMR = datAMR[,AMR],
      imm = datIMM[ ,imm],
   covar
)
    mod = glm(smoking_status~ ., data = df_model, family = binomial(link = "logit"))
    summary(mod)
    part_imm_sign = summary(mod)$coefficients[2:3,4] # check if imm is significant
    
    num_res[[pair]] = c( part_surv_sign,
                        part_imm_sign)
    names(num_res[[pair]]) = c("part_IMM_sign_to_surv", 
                               "part_AMR_sign_to_surv_when_IMM_controlled",
                               "part_AMR_sign_to_smoking",
                               "part_IMM_sign_to_smoking_when_AMR_controlled")
   } else {
     print(paste0(" AMR:", AMR, " and Immune:", imm, " is not kept for doanwstream analysis. "))
     #num_res[[pair]] = c(NA, NA, NA, NA)
     #names(num_res[[pair]]) = c("part_IMM_sign_to_surv", 
    #                            "part_AMR_sign_to_surv_when_IMM_controlled",
     #                           "part_AMR_sign_to_smoking",
     #                           "part_IMM_sign_to_smoking_when_AMR_controlled")
   }

  }


}
num_res = do.call(rbind, num_res)
num_res=data.frame(num_res)

# Directed acyclic graph 1
res_CAT1 = num_res[  num_res$part_IMM_sign_to_surv < pval_thres &
            num_res$part_AMR_sign_to_surv_when_IMM_controlled < pval_thres, ]
res_CAT1
rownames(res_CAT1)
saveRDS(rownames(res_CAT1), file = "results/05_DAG_validated_CAT1_withLF.rds")

# Directed acyclic graph 2
res_CAT2 = num_res[  num_res$part_IMM_sign_to_surv < pval_thres &
            num_res$part_AMR_sign_to_surv_when_IMM_controlled >= pval_thres, ]
rownames(res_CAT2)
saveRDS(rownames(res_CAT2), file = "results/05_DAG_validated_CAT2_withLF.rds")

```


```{r}
sessionInfo()
```
