==> real_data/06_TCGA_serial_mediation.Rmd <==
---
title: "Serial mediation"
output:
  html_document:
    toc: true
    theme: united
date: "Last compiled on `r format(Sys.time(), '%d %B, %Y')`"
---


We sought to determine whether indirect effects exist between tobacco exposure and survival in PDAC patients through mediator chains (AMR–immune infiltration pairs identified by causal discovery), acknowledging that in the context of multiple mediators the total effect can be decomposed by considering them simultaneously.




## Mediation analysis for each AMR - Immune cells type  pairs selected from causal discovery




```{r}
source("surv_serial_mediation_V2.R")

library(dplyr)
set.seed(123)


med_tobacco_AMR = readRDS("results/03_tcga_top50_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")

mean_meth_AMR = readRDS("results/03_tcga_mean_meth_AMR_fdr0_05_V2_K8_corrected.rds")
 
res_med = readRDS("results/02_tcga_med_tobacco_dnam_V2_K8_corrected.rds")

tcga_data = readRDS("results/01_tcga_data_expo_deconv.rds")
X_deconv = tcga_data$X_deconv$Consensus
immune_cells = c("Macrophages","B cells", "T cells", "NK","DCs")
X_immune = X_deconv[rownames(X_deconv) %in% immune_cells,]
immune_cons = t(X_immune)
immune_cons = as.data.frame(immune_cons)
AMRs = colnames(mean_meth_AMR)
mean_meth_AMR = as.data.frame(mean_meth_AMR)
expo = res_med$hdmax2_step1_param$input$exposure_input
covar = res_med$hdmax2_step1_param$input$covar
time = res_med$hdmax2_step1_param$input$survival_time_input
status = res_med$hdmax2_step1_param$input$censoring_status_input

#LF 
lfA = res_med$hdmax2_step1_param$AS_1$U[,1] 
lfB = res_med$hdmax2_step1_param$AS_1$U[,2]
lfC = res_med$hdmax2_step1_param$AS_1$U[,3]
lfD = res_med$hdmax2_step1_param$AS_1$U[,4]
lfE = res_med$hdmax2_step1_param$AS_1$U[,5]
lfF = res_med$hdmax2_step1_param$AS_1$U[,6]
lfG = res_med$hdmax2_step1_param$AS_1$U[,7]
lfH = res_med$hdmax2_step1_param$AS_1$U[,8]

# pairs list
# DAGvalidated_CAT1 <- readRDS("results/05_DAG_validated_CAT1.rds")
# DAGvalidated_CAT1 <- gsub("_", " ", DAGvalidated_CAT1)
# 
# DAGvalidated_CAT2 <- readRDS("results/05_DAG_validated_CAT2.rds")
# DAGvalidated_CAT2 <- gsub("_", " ", DAGvalidated_CAT2)

select_LF = readRDS("results/05_signLFs_by_pairs-A-I.rds")

covars = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3])

# create total immune class 
immune_cons$all = rowSums(immune_cons)




DAGvalidated_CAT1 <- readRDS("results/05_DAG_validated_CAT1_withLF_2.rds")
DAGvalidated_CAT1 <- gsub("_", " ", DAGvalidated_CAT1)



# AMR 4 BC
# error
# AMR 4 all

mean_meth_AMR_temp = mean_meth_AMR$AMR4
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_init <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)

res_mediation_init$feat = c(rep("AMR4_B cells",5))

res = as.data.frame(res_mediation_init)

# AMR 5 B cells

mean_meth_AMR_temp = mean_meth_AMR$AMR5
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR5_B cells",5))
res = rbind(res, res_mediation_temp)

# AMR 5 T cells

mean_meth_AMR_temp = mean_meth_AMR$AMR5
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfB)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR5_T cells",5))
res = rbind(res, res_mediation_temp)

# AMR 9 T cells

mean_meth_AMR_temp = mean_meth_AMR$AMR9
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR9_T cells",5))
res = rbind(res, res_mediation_temp)




# AMR 9  all

mean_meth_AMR_temp = mean_meth_AMR$AMR9
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR9_all",5))
res = rbind(res, res_mediation_temp)

# AMR 11 bc

mean_meth_AMR_temp = mean_meth_AMR$AMR11
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR11_B cells",5))
res = rbind(res, res_mediation_temp)


# AMR 11 bc

mean_meth_AMR_temp = mean_meth_AMR$AMR11
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR11_T cells",5))
res = rbind(res, res_mediation_temp)

# # AMR 11 all
# mean_meth_AMR_temp = mean_meth_AMR$AMR11
# mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)
# 
# immune_cons_temp = immune_cons$all
# immune_cons_temp = as.matrix(immune_cons_temp)
# 
# covars_LF = covars
# 
# res_mediation_temp <- run_serial_mediation_grid2_alter(
#   mat_dmr = mean_meth_AMR_temp,
#   mat_cell = immune_cons_temp,
#   expo = expo,
#   covars_LF = covars_LF,
#   covars = covars,
#   time = time,
#   status = status,
#   R = 1000  # pour aller plus vite
# )
# res_mediation_temp$feat = c(rep("AMR11_all",5))
# res = rbind(res, res_mediation_temp)

## error

# AMR 12 BC 

mean_meth_AMR_temp = mean_meth_AMR$AMR12
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR12_B cells",5))
res = rbind(res, res_mediation_temp)


# AMR 15 all 

mean_meth_AMR_temp = mean_meth_AMR$AMR15
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA, lfB, lfH)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR15_all",5))
res = rbind(res, res_mediation_temp)


# AMR 17 TC 

mean_meth_AMR_temp = mean_meth_AMR$AMR17
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfB)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR17_T cells",5))
res = rbind(res, res_mediation_temp)


# AMR 17 all 

mean_meth_AMR_temp = mean_meth_AMR$AMR17
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA, lfB)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR17_all",5))
res = rbind(res, res_mediation_temp)



# AMR 19 TC 

mean_meth_AMR_temp = mean_meth_AMR$AMR19
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfB)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR19_T cells",5))
res = rbind(res, res_mediation_temp)






# AMR 20 TC 

mean_meth_AMR_temp = mean_meth_AMR$AMR20
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR20_T cells",5))
res = rbind(res, res_mediation_temp)



# AMR 20 all 

mean_meth_AMR_temp = mean_meth_AMR$AMR20
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR20_all",5))
res = rbind(res, res_mediation_temp)


# AMR 23 BC

mean_meth_AMR_temp = mean_meth_AMR$AMR23
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR23_B cells",5))
res = rbind(res, res_mediation_temp)

# AMR 23 TC

mean_meth_AMR_temp = mean_meth_AMR$AMR23
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR23_T cells",5))
res = rbind(res, res_mediation_temp)


# AMR 23 all

mean_meth_AMR_temp = mean_meth_AMR$AMR23
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR23_all",5))
res = rbind(res, res_mediation_temp)

# AMR 27 BC

mean_meth_AMR_temp = mean_meth_AMR$AMR27
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR27_B cells",5))
res = rbind(res, res_mediation_temp)

# AMR 27 TC

mean_meth_AMR_temp = mean_meth_AMR$AMR27
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR27_T cells",5))
res = rbind(res, res_mediation_temp)


# AMR 27 all

mean_meth_AMR_temp = mean_meth_AMR$AMR27
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR27_all",5))
res = rbind(res, res_mediation_temp)




# AMR 36 BC

mean_meth_AMR_temp = mean_meth_AMR$AMR36
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR36_B cells",5))
res = rbind(res, res_mediation_temp)

# AMR 23 TC

mean_meth_AMR_temp = mean_meth_AMR$AMR36
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = covars

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR36_T cells",5))
res = rbind(res, res_mediation_temp)


# AMR 36 all

mean_meth_AMR_temp = mean_meth_AMR$AMR36
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR36_all",5))
res = rbind(res, res_mediation_temp)


# AMR 38 BC

mean_meth_AMR_temp = mean_meth_AMR$AMR38
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`B cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfF)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR38_B cells",5))
res = rbind(res, res_mediation_temp)

# AMR 38 TC

mean_meth_AMR_temp = mean_meth_AMR$AMR38
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$`T cells`
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfB,lfF)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR38_T cells",5))
res = rbind(res, res_mediation_temp)


# AMR 38 all

mean_meth_AMR_temp = mean_meth_AMR$AMR38
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA,lfB,lfF)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR38_all",5))
res = rbind(res, res_mediation_temp)

# AMR 48 all

mean_meth_AMR_temp = mean_meth_AMR$AMR48
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA,lfF)

res_mediation_temp <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)
res_mediation_temp$feat = c(rep("AMR48_all",5))
res = rbind(res, res_mediation_temp)



colnames(res) = c("estimate","ci_low","ci_high", "effect_type", "feat")

saveRDS(res, "results/06_TCGA_CAT1_serial_med_V2_K8_corrected_with_selected_LF.rds")
```


```{r}

########
# DAG 2 
########


source("surv_serial_mediation_V2.R")

library(dplyr)
set.seed(123)


med_tobacco_AMR = readRDS("results/03_tcga_top50_tobacco_AMR_fdr0_05_V2_K8_corrected.rds")

mean_meth_AMR = readRDS("results/03_tcga_mean_meth_AMR_fdr0_05_V2_K8_corrected.rds")
 
res_med = readRDS("results/02_tcga_med_tobacco_dnam_V2_K8_corrected.rds")

tcga_data = readRDS("results/01_tcga_data_expo_deconv.rds")
X_deconv = tcga_data$X_deconv$Consensus
immune_cells = c("Macrophages","B cells", "T cells", "NK","DCs")
X_immune = X_deconv[rownames(X_deconv) %in% immune_cells,]
immune_cons = t(X_immune)
immune_cons = as.data.frame(immune_cons)
AMRs = colnames(mean_meth_AMR)
mean_meth_AMR = as.data.frame(mean_meth_AMR)
expo = res_med$hdmax2_step1_param$input$exposure_input
covar = res_med$hdmax2_step1_param$input$covar
time = res_med$hdmax2_step1_param$input$survival_time_input
status = res_med$hdmax2_step1_param$input$censoring_status_input

#LF 
lfA = res_med$hdmax2_step1_param$AS_1$U[,1] 
lfB = res_med$hdmax2_step1_param$AS_1$U[,2]
lfC = res_med$hdmax2_step1_param$AS_1$U[,3]
lfD = res_med$hdmax2_step1_param$AS_1$U[,4]
lfE = res_med$hdmax2_step1_param$AS_1$U[,5]
lfF = res_med$hdmax2_step1_param$AS_1$U[,6]
lfG = res_med$hdmax2_step1_param$AS_1$U[,7]
lfH = res_med$hdmax2_step1_param$AS_1$U[,8]

# pairs list
# DAGvalidated_CAT1 <- readRDS("results/05_DAG_validated_CAT1.rds")
# DAGvalidated_CAT1 <- gsub("_", " ", DAGvalidated_CAT1)
# 
# DAGvalidated_CAT2 <- readRDS("results/05_DAG_validated_CAT2.rds")
# DAGvalidated_CAT2 <- gsub("_", " ", DAGvalidated_CAT2)

select_LF = readRDS("results/05_signLFs_by_pairs-A-I.rds")

covars = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3])

# create total immune class 
immune_cons$all = rowSums(immune_cons)


set.seed(456)
DAGvalidated_CAT2 <- readRDS("results/05_DAG_validated_CAT2_withLF_2.rds")
DAGvalidated_CAT2 <- gsub("_", " ", DAGvalidated_CAT2)

mean_meth_AMR_temp = mean_meth_AMR$AMR7
mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)

immune_cons_temp = immune_cons$all
immune_cons_temp = as.matrix(immune_cons_temp)

covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA)

res_mediation_init <- run_serial_mediation_grid2_alter(
  mat_dmr = mean_meth_AMR_temp,
  mat_cell = immune_cons_temp,
  expo = expo,
  covars_LF = covars_LF,
  covars = covars,
  time = time,
  status = status,
  R = 1000  # pour aller plus vite
)

res_mediation_init$feat = c(rep("AMR7_all",5))

res = as.data.frame(res_mediation_init)

# AMR 8 all

# mean_meth_AMR_temp = mean_meth_AMR$AMR8
# mean_meth_AMR_temp = as.matrix(mean_meth_AMR_temp)
# 
# immune_cons_temp = immune_cons$all
# immune_cons_temp = as.matrix(immune_cons_temp)
# 
# covars_LF = data.frame(grade = covar[,1], age = covar[,2], gender = covar[,3], lfA)
# 
# res_mediation_temp <- run_serial_mediation_grid2_alter(
#   mat_dmr = mean_meth_AMR_temp,
#   mat_cell = immune_cons_temp,
#   expo = expo,
#   covars_LF = covars_LF,
#   covars = covars,
#   time = time,
#   status = status,
#   R = 10000  # pour aller plus vite
# )
# res_mediation_temp$feat = c(rep("AMR8_all",5))
# res = rbind(res, res_mediation_temp)




colnames(res) = c("estimate","ci_low","ci_high", "effect_type", "feat")

saveRDS(res, "results/06_TCGA_CAT2_serial_med_V2_K8_corrected_with_selected_LF.rds")

```
